improved!

+++ controllers/posts.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import PostModel, { postObject } from "../models/Post";
 
 export const getPosts = async (
-  req: Request,
+  req: Request | any,
   res: Response,
   next: NextFunction
 ): Promise<void> => {
@@ -20,7 +20,6 @@ export const getPostById = async (
   next: NextFunction
 ): Promise<any> => {
   const id = req.params.id;
-  console.log(typeof id);
   try {
     const post = await PostModel.findById(id);
     res.status(200).json(post);
@@ -34,12 +33,12 @@ export const getPostById = async (
 };
 
 export const createPost = async (
-  req: Request,
+  req: Request | any,
   res: Response,
   next: NextFunction
 ): Promise<void> => {
   try {
-    let newPost = new PostModel({ ...req.body });
+    let newPost = new PostModel({ ...req.body, creator: req!.userId });
     await newPost.save();
     res.status(201).json(newPost);
   } catch (error) {
@@ -49,18 +48,26 @@ export const createPost = async (
 };
 
 export const updatePostById = async (
-  req: Request,
+  req: Request | any,
   res: Response,
   next: NextFunction
 ): Promise<any> => {
   const id = req.params.id;
 
   try {
+    const userPost = await PostModel.find({ creator: req!.userId });
+    const index = userPost.findIndex((el) => el._id === id.toString());
+    if (index !== -1) {
+      return res
+        .status(400)
+        .json({ msg: "post can only be edited by owned user only!" });
+    }
     const updatedPost = await PostModel.findByIdAndUpdate(
       id,
       { $set: { ...req.body, _id: id } },
       { new: true }
     );
+
     res.status(200).json(updatedPost);
   } catch (error) {
     if (error.kind === "ObjectId") {
@@ -72,12 +79,23 @@ export const updatePostById = async (
 };
 
 export const deletePostById = async (
-  req: Request,
+  req: Request | any,
   res: Response,
   next: NextFunction
 ): Promise<any> => {
   const id = req.params.id;
+  console.log(id);
+  console.log(req!.userId);
   try {
+    const userPost = await PostModel.find({ creator: req!.userId });
+
+    const index = userPost.findIndex((el) => el._id === id.toString());
+    console.log(index);
+    if (index !== -1) {
+      return res
+        .status(400)
+        .json({ msg: "post can only be deleted by owned user only!" });
+    }
     await PostModel.findByIdAndRemove(id);
     res.status(204).json({ message: "deleted successfully" });
   } catch (error) {
@@ -90,14 +108,23 @@ export const deletePostById = async (
 };
 
 export const likePostOfId = async (
-  req: Request,
+  req: Request | any,
   res: Response,
   next: NextFunction
 ): Promise<any> => {
   const id = req.params.id;
   try {
-    let post = await PostModel.findOne({ _id: id });
-    post!.likeCount = post!.likeCount + 1;
+    let post = await PostModel.findById(id);
+    const index = post!.likeCount.findIndex(
+      (el) => el === req!.userId.toString()
+    );
+    if (index === -1) {
+      post!.likeCount!.push(req.userId);
+    } else {
+      post!.likeCount = post!.likeCount.filter(
+        (el) => el.toString() !== req!.userId.toString()
+      );
+    }
     await post!.save();
     res.status(200).json(post);
   } catch (error) {
